+++ src/services/CasperServiceByProvider.ts
@@ -1,11 +1,12 @@
 import Client, { RequestManager } from '@open-rpc/client-js';
 import ProviderTransport, {
   SafeEventEmitterProvider
-} from '../lib/ProviderTransport';
+} from './ProviderTransport';
 import { CasperServiceByJsonRPC } from './CasperServiceByJsonRPC';
 
 export class CasperServiceByProvider extends CasperServiceByJsonRPC {
   constructor(provider: SafeEventEmitterProvider) {
+    // Hack to override base class
     super('https://example.com');
     const transport = new ProviderTransport(provider);
     const requestManager = new RequestManager([transport]);
